scripts/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
